Pulsing Syntax Improvement
Instead of sheduling a removal of the class just before the pulse, you can simply write those two functions one after another. The addClass function will only be called after the removeClass function. 

I haven't tested it, so I might be wrong, but I think the time it takes for the removeClass function to remove the class is so tiny that you won't see any delays in the scaling of the player.

Great video by the way, keep on hacking!